+++ app/controllers/accounts_controller.rb
@@ -3,18 +3,17 @@
 # Allows users to modify and edit existing accounts.
 class AccountsController < ApplicationController
   before_action :authenticate_user!
-  before_action :set_account, only: [:show, :edit, :update, :destroy]
-  before_action :redirect_without_account, only: [ :show, :edit, :update, :destroy ]
+  before_action :find_account_or_redirect, only: [:show, :edit, :update, :destroy]
 
   # GET /accounts
-  # GET /accounts.json
   def index
-    @order = scrub_order(Account, params[:order], 'accounts.name')
-    @accounts = current_user.accounts.search(params[:search]).order(@order).page(params[:page]).per( 40 )
+    @order = scrub_order(Account, params[:order], 'accounts.archived, accounts.name')
+    @accounts = current_user.accounts.search(params[:search])
+                            .order(@order)
+                            .page(params[:page]).per(40)
   end
 
   # GET /accounts/1
-  # GET /accounts/1.json
   def show
   end
 
@@ -28,50 +27,35 @@ class AccountsController < ApplicationController
   end
 
   # POST /accounts
-  # POST /accounts.json
   def create
     @account = current_user.accounts.new(account_params)
-
-    respond_to do |format|
-      if @account.save
-        format.html { redirect_to @account, notice: 'Account was successfully created.' }
-        format.json { render action: 'show', status: :created, location: @account }
-      else
-        format.html { render action: 'new' }
-        format.json { render json: @account.errors, status: :unprocessable_entity }
-      end
+    if @account.save
+      redirect_to @account, notice: 'Account was successfully created.'
+    else
+      render :new
     end
   end
 
-  # PUT /accounts/1
-  # PUT /accounts/1.json
+  # PATCH /accounts/1
   def update
-    respond_to do |format|
-      if @account.update(account_params)
-        format.html { redirect_to @account, notice: 'Account was successfully updated.' }
-        format.json { head :no_content }
-      else
-        format.html { render action: 'edit' }
-        format.json { render json: @account.errors, status: :unprocessable_entity }
-      end
+    if @account.update(account_params)
+      redirect_to @account, notice: 'Account was successfully updated.'
+    else
+      render :edit
     end
   end
 
   # DELETE /accounts/1
-  # DELETE /accounts/1.json
   def destroy
     @account.destroy
-
-    respond_to do |format|
-      format.html { redirect_to accounts_path }
-      format.json { head :no_content }
-    end
+    redirect_to accounts_path
   end
 
   private
 
-  def set_account
+  def find_account_or_redirect
     @account = current_user.accounts.find_by_id params[:id]
+    redirect_without_account
   end
 
   def redirect_without_account
@@ -79,6 +63,6 @@ class AccountsController < ApplicationController
   end
 
   def account_params
-    params.require(:account).permit(:name)
+    params.require(:account).permit(:name, :archived)
   end
 end