+++ db/migrate/20160815021314_add_archived_to_accounts.rb
@@ -1,0 +1,6 @@
+class AddArchivedToAccounts < ActiveRecord::Migration
+  def change
+    add_column :accounts, :archived, :boolean, null: false, default: false
+    add_index :accounts, :archived
+  end
+end